вроде работает но криво
