+++ src/Main.java
@@ -2,8 +2,10 @@
 
 public class Main {
     public static void main(String[] args) {
-        ChooseCharacter chooseCharacter = new ChooseCharacter();
-        chooseCharacter.chooseCharacter();
+        ChooseCharacter1 ChooseCharacter1 = new ChooseCharacter1();
+        ChooseCharacter1.chooseCharacter1();
+        ChooseCharacter2 ChooseCharacter2 = new ChooseCharacter2();
+        ChooseCharacter2.chooseCharacter2();
         BattleMechanic battleMechanic = new BattleMechanic();
         battleMechanic.battle();
